Day 3 + switch to unsigned int 64

+++ AoC 2020.tests/Challenge1.cs
@@ -1,3 +1,4 @@
+using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace AoC_2020.tests
@@ -28,14 +29,14 @@ namespace AoC_2020.tests
         public void TestTask1()
         {
             var output = c.task1();
-            Assert.AreEqual(514579, output);
+            Assert.AreEqual((ulong)514579, output);
         }
 
         [TestMethod]
         public void TestTask2()
         {
             var output = c.task2();
-            Assert.AreEqual(241861950, output);
+            Assert.AreEqual((ulong)241861950, output);
         }
     }
 }
